feat: Add will-resize event to TopLevelWindow
The new `will-resize` event can be used to override the new window size
(or even prevent it from changing) before the native window is actually
resized.  This is in contrast to the existing the `resize` event, which
is sent *after* the native window has already been resized. For apps
with e.g. custom window snapping logic, the `resize` event is not
sufficient because it will result in flickering between the dragged size
and the snapped size.

`will-resize` is only emitted on macOS and Windows.